third_party/WebKit/LayoutTests/fast/events/window-onerror-11-expected.txt

This test should trigger 'window.onerror' for the syntax error in the attribute handler, regardless of how it's set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.onerror: "Uncaught SyntaxError: Unexpected token %" at window-onerror-11.html (Line: 11, Column: 25)
No stack trace.
Returning 'true': the error should not be reported in the console as an unhandled exception.



window.onerror: "Uncaught SyntaxError: Unexpected token %" at window-onerror-11.html (Line: 33, Column: 14)
No stack trace.
Returning 'true': the error should not be reported in the console as an unhandled exception.



window.onerror: "Uncaught SyntaxError: Unexpected token %" at window-onerror-11.html (Line: 33, Column: 14)
No stack trace.
Returning 'true': the error should not be reported in the console as an unhandled exception.



PASS successfullyParsed is true

TEST COMPLETE
Button.  Button 2 Button 3
